+++ quetzal/os/parallel_call.py
@@ -1,6 +1,5 @@
 import json
 import os
-import sys
 import shutil
 import time
 import uuid
@@ -14,18 +13,14 @@ from nbconvert import PythonExporter
 def add_freeze_support(file):
     with open(file, 'r') as pyfile:
         lines = pyfile.readlines()
-    top = [
-        "def main():\n"
-    ]
+    top = ['def main():\n']
     bottom = [
-        "from multiprocessing import freeze_support\n"
-        "if __name__ == '__main__':\n"
-        "    freeze_support()\n"
-        "    main()\n"
+        "from multiprocessing import freeze_support\nif __name__ == '__main__':\n    freeze_support()\n    main()\n"
     ]
     lines = top + ['    ' + line for line in lines] + bottom
     with open(file, 'w') as pyfile:
         pyfile.writelines(lines)
+
 
 def split_cwd_and_file(path):
     path_list = path.split('/')
@@ -37,7 +32,7 @@ def split_cwd_and_file(path):
     return cwd, path_list[-1]
 
 
-def parallel_call_jobs(jobs, mode='w', leave=False, workers=1, sleep=1):
+def parallel_call_jobs(jobs, mode='w', leave=False, workers=1, sleep=1, raise_errors=False):
     popens = {}
     for i, file, arg, stdout_file, stderr_file in jobs:
         cwd, file = split_cwd_and_file(file)
@@ -45,17 +40,17 @@ def parallel_call_jobs(jobs, mode='w', leave=False, workers=1, sleep=1):
         with open(stdout_file, mode) as stdout:
             with open(stderr_file, mode) as stderr:
                 if type(arg) == tuple:
-                    command_list = ['python', file] +list(arg)
+                    command_list = ['python', file] + list(arg)
                 else:
-                    command_list = ['python', file] +[arg]
-                #my_env = os.environ
-                #my_env["PYTHONPATH"] = os.pathsep.join(sys.path)[1:]
+                    command_list = ['python', file] + [arg]
+                # my_env = os.environ
+                # my_env["PYTHONPATH"] = os.pathsep.join(sys.path)[1:]
                 popens[i] = Popen(
                     command_list,
                     stdout=stdout,
                     stderr=stderr,
-                    #env=my_env,
-                    cwd=cwd
+                    # env=my_env,
+                    cwd=cwd,
                 )
                 if (i + 1) % workers == 0 or (i + 1) == len(jobs):
                     # print('waiting')
@@ -69,13 +64,15 @@ def parallel_call_jobs(jobs, mode='w', leave=False, workers=1, sleep=1):
         subprocess_name = str(file) + str(i)
         if not leave:
             try:
-                os.remove(os.path.join(cwd,file))
+                os.remove(os.path.join(cwd, file))
             except FileNotFoundError:
                 pass
         with open(stderr_file, 'r', encoding='latin') as stderr:
             content = stderr.read()
-            if 'Error' in content and "end_of_notebook" not in content:
-                print("subprocess **{} {}** terminated with an error.".format(subprocess_name, arg))
+            if 'Error' in content and 'end_of_notebook' not in content:
+                print('subprocess **{} {}** terminated with an error.'.format(subprocess_name, arg))
+                if raise_errors:
+                    raise Exception(content)
 
 
 def parallel_call_notebooks(
@@ -88,7 +85,8 @@ def parallel_call_notebooks(
     leave=False,
     errout_suffix=False,
     freeze_support=True,
-    return_jobs=False
+    return_jobs=False,
+    raise_errors=False,
 ):
 
     start = time.time()
@@ -99,15 +97,14 @@ def parallel_call_notebooks(
         files.append(file)
         if not os.path.exists(file):
             os.system('jupyter nbconvert --to python %s' % notebook)
-        if not os.path.exists(file): # jupyter nb convert failed, for instance, jupyter is not recognized
+        if not os.path.exists(file):  # jupyter nb convert failed, for instance, jupyter is not recognized
             convertNotebook(notebook, file)
             if freeze_support:
                 add_freeze_support(file)
-        
 
     mode = 'w' if errout_suffix else 'a+'
 
-    supported_characters = string.ascii_lowercase + string.ascii_uppercase + string.digits + '-_' 
+    supported_characters = string.ascii_lowercase + string.ascii_uppercase + string.digits + '-_'
     for i in range(len(arg_list)):
         arg = arg_list[i]
         suffix = ''
@@ -123,23 +120,24 @@ def parallel_call_notebooks(
         stderr_file = stderr_path.replace('.txt', '_' + suffix + '.txt')
         jobs.append([i, files[i], arg, stdout_file, stderr_file])
 
-    parallel_call_jobs(jobs, mode=mode, leave=leave, workers=workers, sleep=sleep)
+    parallel_call_jobs(jobs, mode=mode, leave=leave, workers=workers, sleep=sleep, raise_errors=raise_errors)
     end = time.time()
     print(int(end - start), 'seconds')
-    if return_jobs: return jobs
-
+    if return_jobs:
+        return jobs
 
 
 def convertNotebook(notebookPath, modulePath):
 
-  with open(notebookPath) as fh:
-    nb = nbformat.reads(fh.read(), nbformat.NO_CONVERT)
+    with open(notebookPath) as fh:
+        nb = nbformat.reads(fh.read(), nbformat.NO_CONVERT)
 
-  exporter = PythonExporter()
-  source, meta = exporter.from_notebook_node(nb)
+    exporter = PythonExporter()
+    source, meta = exporter.from_notebook_node(nb)
 
-  with open(modulePath, 'w+') as fh:
-    fh.writelines(source)
+    with open(modulePath, 'w+') as fh:
+        fh.writelines(source)
+
 
 def parallel_call_notebook(
     notebook,
@@ -151,21 +149,22 @@ def parallel_call_notebook(
     leave=False,
     errout_suffix=False,
     freeze_support=True,
-    return_jobs=False
+    return_jobs=False,
+    raise_errors=False,
 ):
 
     start = time.time()
     jobs = []
     os.system('jupyter nbconvert --to python %s' % notebook)
     file = notebook.replace('.ipynb', '.py')
-    if not os.path.exists(file): # jupyter nb convert failed, for instance, jupyter is not recognized
+    if not os.path.exists(file):  # jupyter nb convert failed, for instance, jupyter is not recognized
         convertNotebook(notebook, file)
     if freeze_support:
         add_freeze_support(file)
 
     mode = 'w' if errout_suffix else 'a+'
 
-    supported_characters = string.ascii_lowercase + string.ascii_uppercase + string.digits + '-_' 
+    supported_characters = string.ascii_lowercase + string.ascii_uppercase + string.digits + '-_'
     for i in range(len(arg_list)):
         arg = arg_list[i]
         suffix = ''
@@ -181,10 +180,11 @@ def parallel_call_notebook(
         stderr_file = stderr_path.replace('.txt', '_' + suffix + '.txt')
         jobs.append([i, file, arg, stdout_file, stderr_file])
 
-    parallel_call_jobs(jobs, mode=mode, leave=leave, workers=workers, sleep=sleep)
+    parallel_call_jobs(jobs, mode=mode, leave=leave, workers=workers, sleep=sleep, raise_errors=raise_errors)
     end = time.time()
     print(int(end - start), 'seconds')
-    if return_jobs: return jobs
+    if return_jobs:
+        return jobs
 
 
 def parallel_call_python(
@@ -208,11 +208,7 @@ def parallel_call_python(
         # print(i, arg)
         with open(stdout_path.replace('.txt', '_' + suffix + '.txt'), mode) as stdout:
             with open(stderr_path.replace('.txt', '_' + suffix + '.txt'), mode) as stderr:
-                popens[i] = Popen(
-                    ['python', file, arg],
-                    stdout=stdout,
-                    stderr=stderr
-                )
+                popens[i] = Popen(['python', file, arg], stdout=stdout, stderr=stderr)
                 if (i + 1) % workers == 0 or (i + 1) == len(arg_list):
                     # print('waiting')
                     for p in popens.values():
@@ -227,19 +223,13 @@ def parallel_call_python(
         mode = 'r'
         with open(stderr_path.replace('.txt', '_' + suffix + '.txt'), mode) as stderr:
             content = stderr.read()
-            if 'Error' in content and "end_of_notebook" not in content:
-                print("subprocess **{} {}** terminated with an error.".format(i, arg))
+            if 'Error' in content and 'end_of_notebook' not in content:
+                print('subprocess **{} {}** terminated with an error.'.format(i, arg))
     return popens
 
 
 def parallel_call_subprocess(
-    subprocess_filepath,
-    kwarg_list,
-    dump_kwargs=json.dump,
-    load_result=json.load,
-    leave=False,
-    *args,
-    **kwargs
+    subprocess_filepath, kwarg_list, dump_kwargs=json.dump, load_result=json.load, leave=False, *args, **kwargs
 ):
     input_files = []
     output_files = []
@@ -276,15 +266,14 @@ def parallel_call_subprocess(
     return results
 
 
-
 from concurrent.futures import ProcessPoolExecutor
 from multiprocessing import Process, Queue, Manager, Pipe
-import sys
+import numpy as np
 
 
-def parallel_executor(function, method='pool', num_workers:int=1, parallel_kwargs:dict={}, **kwargs):
-    '''
-    function: 
+def parallel_executor(function, method='pool', num_workers: int = 1, parallel_kwargs: dict = {}, **kwargs):
+    """
+    function:
         function to paralellize
     method = 'pool' | 'queue' | 'manager' | 'pipe':
         Pool, queue and pipe mostly the same performance.
@@ -294,40 +283,36 @@ def parallel_executor(function, method='pool', num_workers:int=1, parallel_kwarg
     parallel_kwargs: dict of arrays
         kwargs in the function to be parallelize. This should be a List.
         ex: dijkstra(indices=parallel_kwargs['indices'][i]) where i is in range(num_workers)
-    '''
+    """
     on_lambda = bool(os.environ.get('AWS_EXECUTION_ENV'))
 
     if on_lambda:
         method = 'pipe'
 
     if method == 'pool':
-        results = process_pool_executor(function=function, 
-                                        num_workers=num_workers, 
-                                        parallel_kwargs=parallel_kwargs, 
-                                        **kwargs)
-        
+        results = process_pool_executor(
+            function=function, num_workers=num_workers, parallel_kwargs=parallel_kwargs, **kwargs
+        )
+
     elif method == 'pipe':
-         results = process_pipe_executor(function=function, 
-                                        num_workers=num_workers, 
-                                        parallel_kwargs=parallel_kwargs, 
-                                        **kwargs)
-                        
-    elif method =='queue':
-        results = process_queue_executor(function=function,
-                                        num_workers=num_workers, 
-                                        parallel_kwargs=parallel_kwargs, 
-                                        **kwargs)
+        results = process_pipe_executor(
+            function=function, num_workers=num_workers, parallel_kwargs=parallel_kwargs, **kwargs
+        )
+
+    elif method == 'queue':
+        results = process_queue_executor(
+            function=function, num_workers=num_workers, parallel_kwargs=parallel_kwargs, **kwargs
+        )
 
     elif method == 'manager':
-        results = process_manager_executor(function=function, 
-                                            num_workers=num_workers, 
-                                            parallel_kwargs=parallel_kwargs, 
-                                            **kwargs)
-        
+        results = process_manager_executor(
+            function=function, num_workers=num_workers, parallel_kwargs=parallel_kwargs, **kwargs
+        )
+
     else:
-        print(method,'is not a valid choice')        
+        print(method, 'is not a valid choice')
         return 0
-        
+
     return results
 
 
@@ -335,19 +320,16 @@ def process_pool_executor(function, num_workers=1, parallel_kwargs={}, **kwargs)
     results = []
     with ProcessPoolExecutor(max_workers=num_workers) as executor:
         for i in range(num_workers):
-            batch_kwargs = {key:item[i] for key,item in parallel_kwargs.items()} 
-            p = executor.submit(
-                function,
-                **batch_kwargs,
-                **kwargs
-            )
+            batch_kwargs = {key: item[i] for key, item in parallel_kwargs.items()}
+            p = executor.submit(function, **batch_kwargs, **kwargs)
             results.append(p)
-    results  = [res.result() for res in results]
+    results = [res.result() for res in results]
     return results
 
+
 def process_wrapper(function, process_kwargs, kwargs, queue, i):
-    result = function(**process_kwargs,**kwargs)
-    queue.put((i,result))
+    result = function(**process_kwargs, **kwargs)
+    queue.put((i, result))
 
 
 def process_queue_executor(function, num_workers=1, parallel_kwargs={}, **kwargs):
@@ -357,21 +339,21 @@ def process_queue_executor(function, num_workers=1, parallel_kwargs={}, **kwargs
     processes = []
     results = [None] * num_workers
     for i in range(num_workers):
-        process_kwargs = {key:item[i] for key,item in parallel_kwargs.items()} 
+        process_kwargs = {key: item[i] for key, item in parallel_kwargs.items()}
         process = Process(target=process_wrapper, args=(function, process_kwargs, kwargs, queue, i))
         process.start()
         processes.append(process)
     for _ in processes:
         res = queue.get()
-        results[res[0]]=res[1] 
+        results[res[0]] = res[1]
     for process in processes:
-        process.join()      
-
+        process.join()
 
     return results
 
+
 def process_wrapper_manager(function, process_kwargs, kwargs, result_list, i):
-    result = function(**process_kwargs,**kwargs)
+    result = function(**process_kwargs, **kwargs)
     result_list[i] = result
 
 
@@ -382,20 +364,20 @@ def process_manager_executor(function, num_workers=1, parallel_kwargs={}, **kwar
     manager = Manager()
     results = manager.list([None] * num_workers)
     for i in range(num_workers):
-        process_kwargs = {key:item[i] for key,item in parallel_kwargs.items()} 
+        process_kwargs = {key: item[i] for key, item in parallel_kwargs.items()}
         process = Process(target=process_wrapper_manager, args=(function, process_kwargs, kwargs, results, i))
         process.start()
         processes.append(process)
     for process in processes:
         process.join()
     # Convert the manager list to a regular list for easier access
-    results = np.array(results, dtype="object")   
+    results = np.array(results, dtype='object')
     return results
 
 
-def process_wrapper_pipe(function, process_kwargs, kwargs, conn,i):
-    result = function(**process_kwargs,**kwargs)
-    conn.send((i,result))
+def process_wrapper_pipe(function, process_kwargs, kwargs, conn, i):
+    result = function(**process_kwargs, **kwargs)
+    conn.send((i, result))
 
 
 def process_pipe_executor(function, num_workers=1, parallel_kwargs={}, **kwargs):
@@ -411,13 +393,13 @@ def process_pipe_executor(function, num_workers=1, parallel_kwargs={}, **kwargs)
         parent_conn, child_conn = Pipe()
         parent_connections.append(parent_conn)
         child_connections.append(child_conn)
-        process_kwargs = {key:item[i] for key,item in parallel_kwargs.items()} 
-        process = Process(target=process_wrapper_pipe, args=(function, process_kwargs, kwargs, child_conn,i))
+        process_kwargs = {key: item[i] for key, item in parallel_kwargs.items()}
+        process = Process(target=process_wrapper_pipe, args=(function, process_kwargs, kwargs, child_conn, i))
         process.start()
         processes.append(process)
     for i in range(num_workers):
         res = parent_connections[i].recv()
-        results[res[0]]=res[1] 
+        results[res[0]] = res[1]
         child_connections[i].close()
     for process in processes:
         process.join()